requirements.txt
asgiref==3.8.1
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.7
Django==5.0.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
django-rest-knox==4.2.0
djangorestframework==3.15.1
djangorestframework-jsonapi==7.0.0
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
pycparser==2.22
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
